--- src/components/ConnectWallet/ConnectWallet.less
@@ -1,8 +1,0 @@
-.connect-wallet {
-  &__connect-btn {
-    width: 175px;
-    border-color: var(--ergo-connect-wallet-btn-border) !important;
-    background: var(--ergo-connect-wallet-btn-bg) !important;
-    color: var(--ergo-connect-wallet-btn-color) !important;
-  }
-}
