+++ src/components/Header/NetworkDropdown/NetworkDropdown.less
@@ -1,4 +1,4 @@
-@import './src/assets/styles/variables/gradients';
+@import '../../../assets/styles/variables/gradients';
 
 .network-dropdown {
   &__btn {
